+++ src/nucleus.cpp
@@ -94,7 +94,7 @@ double Nucleus::nuclear_thickness_optical(double x,double y){
 		int NZ= int(2*ZMaxAbs/dz);
 
 		double thickness_unnorm=0;
-		for (size_t iz = 0; iz < NZ; iz++) {
+		for (int iz = 0; iz < NZ; iz++) {
 			double zz = dz*iz - ZMaxAbs;
 			if(iz==0 or iz==NZ-1){CC=1/2.;}
 			else{CC=1.;}
